pyproject.toml
[project]
name = "merktrouw"
version = "0.1.4"
description = "Merktrouw project"
requires-python = ">=3.11, <3.12"

dependencies = [
    "mlflow==2.15.1", # Used to be 2.17.0
    "cffi==1.17.1",
    "cloudpickle==3.1.0",
    "matplotlib==3.9.2",
    "numpy<1.25,>=1.21",  # Downgrade for compatibility
    "pandas<2.1,>1.1",  # Downgrade for compatibility
    "psutil==6.0.0",
    "pyarrow==14.0.1",
    "scikit-learn==1.5.2",
    "xgboost==2.1.3",
    "scipy<1.12,>=1.4.1",  # Downgrade for compatibility
    "databricks-feature-engineering==0.6",
    "databricks-feature-lookup==1.2.0",
    "databricks-sdk==0.32.0",
    "pydantic<2",  # Downgrade for compatibility
    "pyspark>=2.1.0",  # Add missing dependency
    "loguru==0.7.3"
]



[project.optional-dependencies]
dev = ["databricks-connect>=15.4.1, <16",
       "ipykernel>=6.29.5, <7",
       "pip>=24.2",
       "pre-commit>=4.1.0, <5"]

[build-system]
requires = [
  "setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where =["src"]
include = ["merktrouw*"]

[tool.ruff]
line-length = 120
select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
